+++ tests/test/test_wolk_connector.c
@@ -30,6 +30,7 @@ TEST_FILE("MQTTSubscribeClient.c")
 #include "actuator_command.h"
 #include "actuator_status.h"
 #include "configuration_command.h"
+#include "utc_command.h"
 
 #include "persistence.h"
 #include "in_memory_persistence.h"
@@ -163,7 +164,7 @@ void test_wolk_connector_wolk_add_string_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     char value[READING_SIZE];
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -183,7 +184,7 @@ void test_wolkconnector_wolk_add_multi_value_string_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     char value[READING_DIMENSIONS][READING_SIZE];
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -205,7 +206,7 @@ void test_wolk_connector_wolk_add_numeric_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     double value = 32.1;
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -224,7 +225,7 @@ void test_wolk_connector_wolk_add_multi_value_numeric_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     double value[READING_SIZE];
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -246,7 +247,7 @@ void test_wolk_connector_wolk_add_bool_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     bool value = 1;
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -265,7 +266,7 @@ void test_wolk_connector_wolk_add_multi_value_bool_sensor_reading(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     bool value[READING_SIZE];
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
 
@@ -287,7 +288,7 @@ void  test_wolk_connector_wolk_add_alarm(void)
     wolk_ctx_t wolk;
     char reference[MANIFEST_ITEM_REFERENCE_SIZE];
     bool value;
-    uint32_t utc_time = 1592574949;
+    uint64_t utc_time = 1592574949;
     uint8_t persistence_storage[1024 * 1024];
     static outbound_message_t outbound_message;
